files

+++ src/js/site/Settings.js
@@ -6,8 +6,8 @@ const fs       = require('fs');
 
 function hasSettings()
 {
-	// return false;
-	return settings.has('path.storage');
+	return false;
+	// return settings.has('path.storage');
 }
 
 function getSettings()
@@ -25,12 +25,18 @@ function setSettings(value)
 function createDefaults()
 {
 	const storagePath = getSettings();
-	fs.mkdirSync(storagePath + '/json/');
-	fs.mkdirSync(storagePath + '/media/');
-	
-	fs.writeFileSync(storagePath + '/json/patch.json', JSON.stringify({}));
-	
-	fs.writeFileSync(storagePath + '/json/data.json', JSON.stringify({
+
+	if (!fs.existsSync(storagePath + '/json/'))
+		fs.mkdirSync(storagePath + '/json/');
+
+	if (!fs.existsSync(storagePath + '/media/'))
+		fs.mkdirSync(storagePath + '/media/');
+
+	if (!fs.existsSync(storagePath + '/json/patch.json'))
+		fs.writeFileSync(storagePath + '/json/patch.json', JSON.stringify({}));
+
+	if (!fs.existsSync(storagePath + '/json/data.json'))
+		fs.writeFileSync(storagePath + '/json/data.json', JSON.stringify({
 		"version" : 0,
 		"tagTypes":
 			[
